+++ lib/foundation/index.ts
@@ -1,2 +1,3 @@
 export * from "./foundation-model";
 export * from "./foundation-view";
+export * from "./relationship-types";